TOP-29 feature signup api (#71)
* modify:TOP-29 Terms 관련 데이터에 key값 추가

 - signup api 보낼때 어떤 요청인지 확인을 위한 값

* modify:TOP-29 SignupResponse 데이터 구조 변경

* feat:TOP-29 SignupUserModel to SignupRequest Mapper 코드 작성

 - deviceKey, regionCode, interestList는 현재 임시 코드이며, 추후 변경 될 예정

* feat:TOP-49 Signup API 적용

* modify:TOP-29 Signup Api 형태 변경

* remove: TOP-29 MockTHTSignupApi 제거, 실제 Api를 연동

* feat:TOP-29 SignupRootViewModel - 회원가입 실패 처리 코드 추가

* fix#29 LocationFragment에서 next event가 rootViewModel로 잘못 전달되고 있던 문제 수정

* feat:TOP-29 SignupRoot에 dataLoading 추가, 회원가입 실패 시 안내 메시지 출력

* modify:TOP-29 Signup Api 주소 변경 적용

+++ data/src/main/java/com/tht/tht/data/local/dao/TokenDao.kt
@@ -6,7 +6,7 @@ interface TokenDao {
 
     fun updateFcmToken(token: String)
 
-    fun updateThtToken(token: String, accessTokenExpiresIn: Int, phone: String)
+    fun updateThtToken(token: String, accessTokenExpiresIn: Long, phone: String)
 
     fun fetchThtToken(): String?
