+++ domain/src/main/java/com/tht/tht/domain/signup/model/SignupResponseModel.kt
@@ -2,5 +2,5 @@ package com.tht.tht.domain.signup.model
 
 data class SignupResponseModel(
     val accessToken: String,
-    val accessTokenExpiresIn: Int
+    val accessTokenExpiresIn: Long
 )
