+++ domain/src/main/java/com/tht/tht/domain/token/model/FcmTokenLoginResponseModel.kt
@@ -2,5 +2,5 @@ package com.tht.tht.domain.token.model
 
 data class FcmTokenLoginResponseModel(
     val accessToken: String,
-    val accessTokenExpiresIn: Int
+    val accessTokenExpiresIn: Long
 )
