+++ domain/src/main/java/com/tht/tht/domain/token/repository/TokenRepository.kt
@@ -6,7 +6,7 @@ interface TokenRepository {
 
     suspend fun updateFcmToken(token: String)
 
-    suspend fun updateThtToken(token: String, accessTokenExpiresIn: Int, phone: String)
+    suspend fun updateThtToken(token: String, accessTokenExpiresIn: Long, phone: String)
 
     suspend fun fetchThtToken(): String?
